pyproject.toml
[project]
name = "asset-search-v2"
version = "0.1.0"
description = "Web-based physical asset discovery pipeline for corporate entities"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.0",
    "pydantic-ai>=1.58.0",
    "psycopg[binary]>=3.2.0",
    "litellm>=1.81.0",
    "instructor>=1.7.0",
    "firecrawl-py>=4.16.0",
    "exa-py>=2.4.0",
    "tiktoken>=0.9.0",
    "rich>=14.0.0",
    "python-dotenv>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asset_search"]
